fixed issue 270; agent installer for x64 version 0.2.31; poller messages improved; added paging to log viewer in java console

+++ src/libnetxms/string.cpp
@@ -49,7 +49,7 @@ String::String(const String &src)
 
 String::String(const TCHAR *init)
 {
-   m_dwBufSize = _tcslen(init) + 1;
+   m_dwBufSize = (DWORD)_tcslen(init) + 1;
    m_pszBuffer = _tcsdup(init);
 }
